feat : #41 area detail페이지 레이아웃 수정
